+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 
 ### Added
 - Maven plugin to generate code from templates ([#347](https://github.com/ehrbase/openEHR_SDK/pull/347))
-- Example Generator ([#349](https://github.com/ehrbase/openEHR_SDK/pull/349))
+- Example Generator ([#349](https://github.com/ehrbase/openEHR_SDK/pull/349), [#351](https://github.com/ehrbase/openEHR_SDK/pull/351))
 
 ### Fixed